requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2-binary==2.9.3
django-extensions==3.1.5
Pillow
drf-yasg==1.20.0
